# Speed up passmanager if it is empty