--- a/qiskit/transpiler/passmanager.py
+++ b/qiskit/transpiler/passmanager.py
@@ -214,12 +214,11 @@ class PassManager:
         """
         if not self._pass_sets and output_name is None and callback is None:
             return circuits
-        elif isinstance(circuits, QuantumCircuit):
+        if isinstance(circuits, QuantumCircuit):
             return self._run_single_circuit(circuits, output_name, callback)
-        elif len(circuits) == 1:
+        if len(circuits) == 1:
             return self._run_single_circuit(circuits[0], output_name, callback)
-        else:
-            return self._run_several_circuits(circuits, output_name, callback)
+        return self._run_several_circuits(circuits, output_name, callback)
 
     def _create_running_passmanager(self) -> RunningPassManager:
         running_passmanager = RunningPassManager(self.max_iteration)